Crate revive_env

This module provides an externalities builder for setting an environment with the Revive runtime.

It is heavily inspired from here: https://github.com/paritytech/revive/blob/56aadce0a9554e68a08feb35cffb65574d7a618b/crates/runner/src/lib.rs#L69

THIS IS WORK IN PROGRESS. It is not yet complete and may change in the future.

Modules

runtime 🔒

Dummy runtime with pallet revive and all necessary pallets for running tests with forge.

Structs

BlockAuthor

ExtBuilder

Externalities builder

GasScale

NativeToEthRatio

Runtime

Enums

RuntimeHoldReason

A reason for placing a hold on funds.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balances

System

Mandatory system pallet that should always be included in a FRAME runtime.

Timestamp

Provides a way for consensus systems to set and check the onchain time.