revive_env

Module runtime

Dummy runtime with pallet revive and all necessary pallets for running tests with forge.

THIS IS WORK IN PROGRESS. It is not yet complete and may change in the future.

Modules

currency

Structs

BlockAuthor

ChainId

CodeHashLockupDepositPercent

DepositPerByte

DepositPerItem

EthExtraImpl

Default extensions applied to Ethereum transactions.

FeeMultiplier

GasScale

NativeToEthRatio

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

TransactionByteFee

UnstableInterface

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

AVERAGE_ON_INITIALIZE_RATIO 🔒

We assume that ~10% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

MAXIMUM_BLOCK_WEIGHT 🔒

We allow for 2 seconds of compute with a 6 second average block time, with maximum proof size.

NORMAL_DISPATCH_RATIO 🔒

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

Traits

InherentDataExt 🔒

Type Aliases

Address

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balances

BalancesConfig

Block

Contracts

ContractsConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

System

Mandatory system pallet that should always be included in a FRAME runtime.

SystemConfig

Timestamp

Provides a way for consensus systems to set and check the onchain time.

TransactionPayment

Provides the ability to charge for extrinsic execution.

TransactionPaymentConfig

TxExtension 🔒

The transaction extensions that are added to the runtime.

UncheckedExtrinsic