revive_env

Module runtime

Dummy runtime with pallet revive and all necessary pallets for running tests with forge.

THIS IS WORK IN PROGRESS. It is not yet complete and may change in the future.

Structs

BlockWeights

CodeHashLockupDepositPercent

DepositPerByte

DepositPerItem

NativeToEthRatio

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

UnstableInterface

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Traits

InherentDataExt 🔒

Type Aliases

AccountId

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balances

BalancesConfig

Block

Contracts

ContractsConfig

System

SystemConfig

Timestamp